Initial import of a remote task service. Here is an excerpt from the README:
This package provides an implementation of a remote task execution Web 
service that allows to execute pre-defined tasks on another server. 
Those services are useful in two ways:

1. They enable us to complete tasks that are not natively available on a
   particular machine. For example, it is not possible to convert an AVI 
   file to a Flash(R) movie using Linux, the operating system our Web 
   server might run on.

2. They also allow to move expensive operations to other servers. This is
   valuable, for example, when converting videos on high-traffic sites.

I think the API is pretty okay, but we will have to gain some experience 
with it first. While I have done some minimal stress testing on my 
machine (mainly ensuring that we do not get ZODB conflicts), it has not 
been used in production yet.

I also have a derivative package specifically for transcoding videos. If 
people are interested in it, I am willing to upload it.



git-svn-id: svn://svn.zope.org/repos/main/lovely.remotetask/trunk@69474 62d5b8a3-27da-0310-9561-8e5933582275